requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Cython==3.0.10
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Cors==5.0.0
fpdf==1.7.2
gunicorn==23.0.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
keyboard==0.13.5
MarkupSafe==3.0.2
numpy==2.1.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==10.4.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.1.3
